+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.0] - 2021-08-19
+### Changed
+- Requires [`httpx`](https://www.python-httpx.org)==0.19.\*
+
 ## [0.10.0] - 2021-04-27
 ### Changed
 - Requires [`httpx`](https://www.python-httpx.org)==0.18.\*
@@ -103,7 +107,8 @@ Note that a few changes were made:
 ### Added
 - Placeholder for port of requests_auth to httpx
 
-[Unreleased]: https://github.com/Colin-b/httpx_auth/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/Colin-b/httpx_auth/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/Colin-b/httpx_auth/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/Colin-b/httpx_auth/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/Colin-b/httpx_auth/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/Colin-b/httpx_auth/compare/v0.7.0...v0.8.0